ruler resize controls
