+++ editor/src/components/canvas/controls/grid-controls.tsx
@@ -17,19 +17,14 @@ import * as EP from '../../../core/shared/element-path'
 import type {
   ElementInstanceMetadataMap,
   GridAutoOrTemplateDimensions,
-  GridContainerProperties,
-  GridPositionOrSpan,
 } from '../../../core/shared/element-template'
 import {
   isGridAutoOrTemplateDimensions,
-  isGridSpan,
   type GridAutoOrTemplateBase,
 } from '../../../core/shared/element-template'
 import type { CanvasPoint, CanvasRectangle, LocalRectangle } from '../../../core/shared/math-utils'
 import {
-  boundingRectangleArray,
   canvasPoint,
-  canvasRectangle,
   isFiniteRectangle,
   isInfinityRectangle,
   nullIfInfinity,
@@ -50,7 +45,6 @@ import { optionalMap } from '../../../core/shared/optional-utils'
 import { assertNever, NO_OP } from '../../../core/shared/utils'
 import { Modifier } from '../../../utils/modifiers'
 import { when } from '../../../utils/react-conditionals'
-import type { UtopiColor } from '../../../uuiui'
 import { useColorTheme, UtopiaStyles } from '../../../uuiui'
 import { useDispatch } from '../../editor/store/dispatch-context'
 import {
@@ -76,21 +70,14 @@ import {
   gridResizeHandle,
 } from '../canvas-strategies/interaction-state'
 import type { GridCellCoordinates } from '../canvas-strategies/strategies/grid-cell-bounds'
-import {
-  getGridChildCellCoordBoundsFromCanvas,
-  gridCellTargetId,
-} from '../canvas-strategies/strategies/grid-cell-bounds'
-import type { GridCellGlobalFrames } from '../canvas-strategies/strategies/grid-helpers'
+import { gridCellTargetId } from '../canvas-strategies/strategies/grid-cell-bounds'
 import {
   getGlobalFrameOfGridCellFromMetadata,
   getGridRelatedIndexes,
   getGridElementPinState,
-  printPin,
   getGridIdentifierContainerOrComponentPath,
   gridIdentifierToString,
   gridIdentifiersSimilar,
-  findOriginalGrid,
-  isAutoGridPin,
 } from '../canvas-strategies/strategies/grid-helpers'
 import { canResizeGridTemplate } from '../canvas-strategies/strategies/resize-grid-strategy'
 import { resizeBoundingBoxFromSide } from '../canvas-strategies/strategies/resize-helpers'
@@ -121,8 +108,7 @@ import { PinOutline, usePropsOrJSXAttributes } from './position-outline'
 import { getLayoutProperty } from '../../../core/layout/getLayoutProperty'
 import { styleStringInArray } from '../../../utils/common-constants'
 import { gridContainerIdentifier, type GridIdentifier } from '../../editor/store/editor-state'
-import type { RulerMarkerType } from './grid-controls-ruler-markers'
-import { rulerMarkerIcons } from './grid-controls-ruler-markers'
+import { RulerMarkers } from './grid-controls-ruler-markers'
 import type { GridData, GridMeasurementHelperData } from './grid-measurements'
 import {
   getGridElementMeasurementHelperData,
@@ -131,7 +117,6 @@ import {
 } from './grid-measurements'
 import type { Property } from 'csstype'
 import { isFeatureEnabled } from '../../../utils/feature-switches'
-import type { ThemeObject } from '../../../uuiui/styles/theme/theme-helpers'
 
 const CELL_ANIMATION_DURATION = 0.15 // seconds
 
@@ -2153,6 +2138,8 @@ function useIsGridItemInteractionActive() {
         store.editor.canvas.interactionSession.activeControl.type === 'GRID_CELL_HANDLE' ||
         // resize (cell)
         store.editor.canvas.interactionSession.activeControl.type === 'GRID_RESIZE_HANDLE' ||
+        // resize (cell, via ruler markers)
+        store.editor.canvas.interactionSession.activeControl.type === 'GRID_RESIZE_RULER_HANDLE' ||
         // resize (abs)
         store.editor.canvas.interactionSession.activeControl.type === 'RESIZE_HANDLE'
       )
@@ -2169,628 +2156,6 @@ function useSelectedGridItems(): ElementPath[] {
   )
 }
 
-export const RulerMarkerIconSize = 11 // px
-
-type RulerMarkerData = {
-  parentGrid: GridContainerProperties
-  cellRect: CanvasRectangle
-  gridRect: CanvasRectangle
-  otherColumnMarkers: Array<RulerMarkerPositionData>
-  otherRowMarkers: Array<RulerMarkerPositionData>
-  columnStart: RulerMarkerPositionData
-  columnEnd: RulerMarkerPositionData
-  rowStart: RulerMarkerPositionData
-  rowEnd: RulerMarkerPositionData
-}
-
-type RulerMarkerPositionData = {
-  markerType: 'selected' | 'target' | 'other'
-  rowOrColumn: 'row' | 'column'
-  top: number
-  left: number
-  position: GridPositionOrSpan | null
-  bound: 'start' | 'end'
-}
-
-interface RulerMarkersProps {
-  setShowGridCellOutlines: (show: boolean) => void
-  path: ElementPath
-}
-
-const RulerMarkers = React.memo((props: RulerMarkersProps) => {
-  const gridRect: CanvasRectangle | null = useEditorState(
-    Substores.metadata,
-    (store) => {
-      const originalGrid = findOriginalGrid(store.editor.jsxMetadata, EP.parentPath(props.path))
-      if (originalGrid == null) {
-        return null
-      }
-
-      return MetadataUtils.getFrameOrZeroRectInCanvasCoords(originalGrid, store.editor.jsxMetadata)
-    },
-    'RulerMarkers gridRect',
-  )
-
-  const parentGridCellGlobalFrames = useEditorState(
-    Substores.metadata,
-    (store) => {
-      const elementMetadata = MetadataUtils.findElementByElementPath(
-        store.editor.jsxMetadata,
-        props.path,
-      )
-      if (elementMetadata == null) {
-        return null
-      }
-
-      return elementMetadata.specialSizeMeasurements.parentGridCellGlobalFrames
-    },
-    'RulerMarkers parentGridCellGlobalFrames',
-  )
-
-  const rulerMarkerData: RulerMarkerData | null = useEditorState(
-    Substores.metadata,
-    (store) => {
-      if (gridRect == null) {
-        return null
-      }
-
-      if (parentGridCellGlobalFrames == null) {
-        return null
-      }
-
-      const elementMetadata = MetadataUtils.findElementByElementPath(
-        store.editor.jsxMetadata,
-        props.path,
-      )
-      if (elementMetadata == null) {
-        return null
-      }
-
-      const elementGridProperties = elementMetadata.specialSizeMeasurements.elementGridProperties
-      if (elementGridProperties == null) {
-        return null
-      }
-
-      const parentGrid = elementMetadata.specialSizeMeasurements.parentContainerGridProperties
-
-      const originalGrid = findOriginalGrid(store.editor.jsxMetadata, EP.parentPath(props.path))
-      if (originalGrid == null) {
-        return null
-      }
-
-      const cellBounds = getGridChildCellCoordBoundsFromCanvas(
-        elementMetadata,
-        parentGridCellGlobalFrames,
-      )
-      if (cellBounds == null) {
-        return null
-      }
-
-      if (parentGridCellGlobalFrames.length === 0) {
-        return null
-      }
-      const firstRow = parentGridCellGlobalFrames[0]
-      const cellBoundsColumnIndex = cellBounds.column - 1
-      const left = firstRow[cellBoundsColumnIndex].x
-      const width = getCellCanvasWidthFromBounds(
-        parentGridCellGlobalFrames,
-        cellBoundsColumnIndex,
-        cellBounds.width,
-      )
-
-      const cellBoundsRowIndex = cellBounds.row - 1
-      if (
-        parentGridCellGlobalFrames.length <= cellBoundsRowIndex ||
-        parentGridCellGlobalFrames[cellBoundsRowIndex].length === 0
-      ) {
-        return null
-      }
-      const firstColumn = parentGridCellGlobalFrames[cellBoundsRowIndex][0]
-      const top = firstColumn.y
-      const height = getCellCanvasHeightFromBounds(
-        parentGridCellGlobalFrames,
-        cellBoundsRowIndex,
-        cellBounds.height,
-      )
-
-      const cellRect = parentGridCellGlobalFrames[cellBounds.row - 1][cellBounds.column - 1]
-      const cellRectResized = canvasRectangle({
-        x: cellRect.x,
-        y: cellRect.y,
-        width: width,
-        height: height,
-      })
-
-      let otherColumnMarkers: Array<RulerMarkerPositionData> = []
-      let otherRowMarkers: Array<RulerMarkerPositionData> = []
-
-      function addOtherMarker(
-        rowOrColumn: RulerMarkerPositionData['rowOrColumn'],
-        bound: RulerMarkerPositionData['bound'],
-        rulerLeft: number,
-        rulerTop: number,
-      ): void {
-        const otherMarker: RulerMarkerPositionData = {
-          markerType: 'other',
-          rowOrColumn: rowOrColumn,
-          top: rulerTop,
-          left: rulerLeft,
-          position: null,
-          bound: bound,
-        }
-        const addTo = rowOrColumn === 'row' ? otherRowMarkers : otherColumnMarkers
-        addTo.push(otherMarker)
-      }
-
-      // Add the additional markers for columns.
-      const lastColumnIndex = parentGridCellGlobalFrames[0].length - 1
-      for (let columnIndex = 0; columnIndex <= lastColumnIndex; columnIndex++) {
-        const cell = parentGridCellGlobalFrames[0][columnIndex]
-        if (left !== cell.x) {
-          addOtherMarker('column', 'start', cell.x, gridRect.y)
-        }
-        if (left + width !== cell.x + cell.width) {
-          addOtherMarker('column', 'end', cell.x + cell.width, gridRect.y)
-        }
-      }
-
-      // Add the additional markers for rows.
-      const lastRowIndex = parentGridCellGlobalFrames.length - 1
-      for (let rowIndex = 0; rowIndex <= lastRowIndex; rowIndex++) {
-        const cell = parentGridCellGlobalFrames[rowIndex][0]
-        if (top !== cell.y) {
-          addOtherMarker('row', 'start', gridRect.x, cell.y)
-        }
-        if (top + height !== cell.y + cell.height) {
-          addOtherMarker('row', 'end', gridRect.x, cell.y + cell.height)
-        }
-      }
-
-      const columnStart: RulerMarkerPositionData = {
-        markerType: 'selected',
-        rowOrColumn: 'column',
-        top: gridRect.y,
-        left: left,
-        position: elementGridProperties.gridColumnStart,
-        bound: 'start',
-      }
-      const columnEnd: RulerMarkerPositionData = {
-        markerType: 'selected',
-        rowOrColumn: 'column',
-        top: gridRect.y,
-        left: left + width,
-        position: elementGridProperties.gridColumnEnd,
-        bound: 'end',
-      }
-      const rowStart: RulerMarkerPositionData = {
-        markerType: 'selected',
-        rowOrColumn: 'row',
-        top: top,
-        left: gridRect.x,
-        position: elementGridProperties.gridRowStart,
-        bound: 'start',
-      }
-      const rowEnd: RulerMarkerPositionData = {
-        markerType: 'selected',
-        rowOrColumn: 'row',
-        top: top + height,
-        left: gridRect.x,
-        position: elementGridProperties.gridRowEnd,
-        bound: 'end',
-      }
-
-      return {
-        parentGrid: parentGrid,
-        cellRect: cellRectResized,
-        gridRect: gridRect,
-        otherColumnMarkers: otherColumnMarkers,
-        otherRowMarkers: otherRowMarkers,
-        columnStart: columnStart,
-        columnEnd: columnEnd,
-        rowStart: rowStart,
-        rowEnd: rowEnd,
-      }
-    },
-    'RulerMarkers markers',
-  )
-
-  const [showExtraMarkers, setShowExtraMarkers] = React.useState<'row' | 'column' | null>(null)
-  const markerMouseUp = React.useCallback(
-    (event: MouseEvent) => {
-      event.preventDefault()
-      event.stopPropagation()
-      setShowExtraMarkers(null)
-      props.setShowGridCellOutlines(false)
-      window.removeEventListener('mouseup', markerMouseUp)
-    },
-    [props],
-  )
-  const rowMarkerMouseDown = React.useCallback(
-    (event: React.MouseEvent) => {
-      event.preventDefault()
-      event.stopPropagation()
-      setShowExtraMarkers('row')
-      props.setShowGridCellOutlines(true)
-      window.addEventListener('mouseup', markerMouseUp)
-    },
-    [markerMouseUp, props],
-  )
-  const columnMarkerMouseDown = React.useCallback(
-    (event: React.MouseEvent) => {
-      event.preventDefault()
-      event.stopPropagation()
-      setShowExtraMarkers('column')
-      props.setShowGridCellOutlines(true)
-      window.addEventListener('mouseup', markerMouseUp)
-    },
-    [markerMouseUp, props],
-  )
-
-  if (rulerMarkerData == null || gridRect == null) {
-    return null
-  }
-
-  return (
-    <React.Fragment>
-      <GridRuler
-        axis='column'
-        gridRect={gridRect}
-        cellFrames={parentGridCellGlobalFrames}
-        rulerVisible={showExtraMarkers === 'column' ? 'visible' : 'not-visible'}
-      >
-        {/* Other markers for unselected tracks */}
-        {rulerMarkerData.otherColumnMarkers.map((marker, index) => {
-          return (
-            <RulerMarkerIndicator
-              key={`ruler-marker-${index}`}
-              gridRect={rulerMarkerData.gridRect}
-              parentGrid={rulerMarkerData.parentGrid}
-              marker={marker}
-              axis={marker.rowOrColumn}
-              testID={`ruler-marker-${marker.rowOrColumn}-${index}`}
-              visible={showExtraMarkers == marker.rowOrColumn ? 'visible' : 'not-visible'}
-            />
-          )
-        })}
-        {/* Selected item markers */}
-        <RulerMarkerIndicator
-          gridRect={rulerMarkerData.gridRect}
-          parentGrid={rulerMarkerData.parentGrid}
-          marker={rulerMarkerData.columnStart}
-          axis={'column'}
-          visible={'visible'}
-          onMouseDown={columnMarkerMouseDown}
-          testID={'ruler-marker-column-start'}
-        />
-        <RulerMarkerIndicator
-          gridRect={rulerMarkerData.gridRect}
-          parentGrid={rulerMarkerData.parentGrid}
-          marker={rulerMarkerData.columnEnd}
-          axis={'column'}
-          visible={'visible'}
-          onMouseDown={columnMarkerMouseDown}
-          testID={'ruler-marker-column-end'}
-        />
-      </GridRuler>
-
-      <GridRuler
-        axis='row'
-        gridRect={gridRect}
-        cellFrames={parentGridCellGlobalFrames}
-        rulerVisible={showExtraMarkers === 'row' ? 'visible' : 'not-visible'}
-      >
-        {/* Other markers for unselected tracks */}
-        {rulerMarkerData.otherRowMarkers.map((marker, index) => {
-          return (
-            <RulerMarkerIndicator
-              key={`ruler-marker-${index}`}
-              gridRect={rulerMarkerData.gridRect}
-              parentGrid={rulerMarkerData.parentGrid}
-              marker={marker}
-              axis={marker.rowOrColumn}
-              testID={`ruler-marker-${marker.rowOrColumn}-${index}`}
-              visible={showExtraMarkers == marker.rowOrColumn ? 'visible' : 'not-visible'}
-            />
-          )
-        })}
-        {/* Selected item markers */}
-        <RulerMarkerIndicator
-          gridRect={rulerMarkerData.gridRect}
-          parentGrid={rulerMarkerData.parentGrid}
-          marker={rulerMarkerData.rowStart}
-          axis={'row'}
-          visible={'visible'}
-          onMouseDown={rowMarkerMouseDown}
-          testID={'ruler-marker-row-start'}
-        />
-        <RulerMarkerIndicator
-          gridRect={rulerMarkerData.gridRect}
-          parentGrid={rulerMarkerData.parentGrid}
-          marker={rulerMarkerData.rowEnd}
-          axis={'row'}
-          visible={'visible'}
-          onMouseDown={rowMarkerMouseDown}
-          testID={'ruler-marker-row-end'}
-        />
-      </GridRuler>
-
-      {/* Offset lines */}
-      <GridCellOffsetLine
-        top={rulerMarkerData.columnStart.top}
-        left={rulerMarkerData.columnStart.left}
-        size={rulerMarkerData.cellRect.y - rulerMarkerData.gridRect.y}
-        orientation='vertical'
-      />
-      <GridCellOffsetLine
-        top={rulerMarkerData.columnEnd.top}
-        left={rulerMarkerData.columnEnd.left}
-        size={rulerMarkerData.cellRect.y - rulerMarkerData.gridRect.y}
-        orientation='vertical'
-      />
-      <GridCellOffsetLine
-        top={rulerMarkerData.rowStart.top}
-        left={rulerMarkerData.rowStart.left}
-        size={rulerMarkerData.cellRect.x - rulerMarkerData.gridRect.x}
-        orientation='horizontal'
-      />
-      <GridCellOffsetLine
-        top={rulerMarkerData.rowEnd.top}
-        left={rulerMarkerData.rowEnd.left}
-        size={rulerMarkerData.cellRect.x - rulerMarkerData.gridRect.x}
-        orientation='horizontal'
-      />
-
-      {/* Cell outline */}
-      <GridCellOutline
-        top={rulerMarkerData.cellRect.y}
-        left={rulerMarkerData.cellRect.x}
-        width={rulerMarkerData.cellRect.width + 1}
-        height={rulerMarkerData.cellRect.height + 1}
-      />
-    </React.Fragment>
-  )
-})
-RulerMarkers.displayName = 'RulerMarkers'
-
-interface GridRulerProps {
-  axis: 'row' | 'column'
-  gridRect: CanvasRectangle
-  cellFrames: GridCellGlobalFrames | null
-  rulerVisible: 'visible' | 'not-visible'
-}
-
-const GridRuler = React.memo<React.PropsWithChildren<GridRulerProps>>(
-  ({ axis, gridRect, rulerVisible, cellFrames, children }) => {
-    const colorTheme = useColorTheme()
-    const scale = useEditorState(
-      Substores.canvas,
-      (store) => store.editor.canvas.scale,
-      'GridRuler scale',
-    )
-
-    // Make sure the ruler extends to cover all the cells and not just the grid dimensions.
-    const cellMaxRect = React.useMemo(() => {
-      return boundingRectangleArray([gridRect, ...(cellFrames?.flat() ?? [])]) ?? gridRect
-    }, [cellFrames, gridRect])
-
-    const columnLeft = gridRect.x
-    const rowLeft = gridRect.x - RulerMarkerIconSize / scale
-    const left = axis === 'row' ? rowLeft : columnLeft
-
-    const columnTop = gridRect.y - RulerMarkerIconSize / scale
-    const rowTop = gridRect.y
-    const top = axis === 'row' ? rowTop : columnTop
-
-    const width = axis === 'row' ? RulerMarkerIconSize / scale : cellMaxRect.width
-    const height = axis === 'row' ? cellMaxRect.height : RulerMarkerIconSize / scale
-
-    return (
-      <div
-        data-testid={`ruler-${axis}`}
-        style={{
-          position: 'absolute',
-          left: left,
-          top: top,
-          width: width,
-          height: height,
-          backgroundColor: rulerVisible === 'visible' ? colorTheme.white.value : 'transparent',
-          outlineWidth: '0.2px',
-          outlineStyle: 'solid',
-          outlineColor: rulerVisible === 'visible' ? colorTheme.black.value : 'transparent',
-        }}
-      >
-        {children}
-      </div>
-    )
-  },
-)
-GridRuler.displayName = 'GridRuler'
-
-const RulerMarkerIndicator = React.memo(
-  (props: {
-    gridRect: CanvasRectangle
-    parentGrid: GridContainerProperties
-    marker: RulerMarkerPositionData
-    axis: 'row' | 'column'
-    testID: string
-    visible: 'visible' | 'not-visible'
-    onMouseDown?: (event: React.MouseEvent) => void
-  }) => {
-    const colorTheme = useColorTheme()
-
-    const canvasScale = useEditorState(
-      Substores.canvasOffset,
-      (store) => store.editor.canvas.scale,
-      'RulerMarkerIndicator canvasScale',
-    )
-
-    const markerType = getRulerMarkerType(props.marker)
-    const markerColor = getRulerMarkerColor(colorTheme, props.marker)
-    const markerIcon = rulerMarkerIcons[markerType][props.axis](markerColor, canvasScale)
-
-    const scaledTop =
-      props.axis === 'row'
-        ? props.marker.top - props.gridRect.y - RulerMarkerIconSize / canvasScale / 2 + 0.5
-        : 0
-
-    const scaledLeft =
-      props.axis === 'column'
-        ? props.marker.left - props.gridRect.x - RulerMarkerIconSize / canvasScale / 2 + 0.5
-        : 0
-
-    const labelText = React.useMemo(() => {
-      if (props.marker.position == null) {
-        return null
-      }
-      return printPin(props.parentGrid, props.marker.position, props.axis)
-    }, [props.marker, props.parentGrid, props.axis])
-
-    const labelClass = 'ruler-marker-label'
-
-    return (
-      <div
-        data-testid={props.testID}
-        style={{
-          position: 'absolute',
-          top: scaledTop,
-          left: scaledLeft,
-          color: colorTheme.primary.value,
-          height: RulerMarkerIconSize / canvasScale,
-          width: RulerMarkerIconSize / canvasScale,
-          display: props.visible === 'visible' ? 'flex' : 'none',
-        }}
-        onMouseDown={props.onMouseDown}
-        css={{
-          [`> .${labelClass}`]: {
-            visibility: 'hidden',
-          },
-          ':hover': {
-            [`> .${labelClass}`]: {
-              visibility: 'visible',
-            },
-          },
-        }}
-      >
-        {markerIcon}
-        {when(
-          labelText != null,
-          <div
-            className={labelClass}
-            style={{
-              position: 'absolute',
-              background: colorTheme.primary.value,
-              borderRadius: 2 / canvasScale,
-              padding: `${3 / canvasScale}px ${6 / canvasScale}px`,
-              color: colorTheme.white.value,
-              height: 20 / canvasScale,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              top: props.axis === 'column' ? -23 / canvasScale : 0,
-              left: props.axis === 'column' ? 0 : undefined,
-              right: props.axis === 'row' ? (RulerMarkerIconSize + 1) / canvasScale : undefined,
-              fontSize: 11 / canvasScale,
-            }}
-          >
-            {labelText}
-          </div>,
-        )}
-      </div>
-    )
-  },
-)
-RulerMarkerIndicator.displayName = 'RulerMarkerIndicator'
-
-function getRulerMarkerType(marker: RulerMarkerPositionData): RulerMarkerType {
-  const isAuto = isAutoGridPin(marker.position)
-  const isSpanStart = marker.bound === 'start' && isGridSpan(marker.position)
-  const isSpanEnd = marker.bound === 'end' && isGridSpan(marker.position)
-
-  if (marker.markerType === 'other' || marker.markerType === 'target') {
-    return 'auto-short'
-  } else if (isSpanStart) {
-    return 'span-start'
-  } else if (isSpanEnd) {
-    return 'span-end'
-  } else if (isAuto) {
-    return 'auto'
-  } else {
-    return 'pinned'
-  }
-}
-
-function getRulerMarkerColor(colorTheme: ThemeObject, marker: RulerMarkerPositionData): UtopiColor {
-  switch (marker.markerType) {
-    case 'selected':
-      return colorTheme.primary
-    case 'target':
-      return colorTheme.brandPurple
-    case 'other':
-      return colorTheme.grey65
-    default:
-      assertNever(marker.markerType)
-  }
-}
-
-function getCellCanvasWidthFromBounds(
-  grid: CanvasRectangle[][],
-  index: number,
-  cells: number,
-): number {
-  if (grid.length === 0) {
-    return 0
-  }
-
-  const currentRow = grid[0]
-  if (currentRow.length <= index) {
-    return 0
-  }
-  if (cells <= 1) {
-    return currentRow[index].width
-  }
-
-  function getPadding() {
-    if (currentRow.length <= 1) {
-      return 0
-    }
-    return currentRow[1].x - (currentRow[0].x + currentRow[0].width)
-  }
-  const padding = getPadding()
-
-  return currentRow.slice(index + 1, index + cells).reduce((acc, curr) => {
-    return acc + curr.width + padding
-  }, currentRow[index].width)
-}
-
-function getCellCanvasHeightFromBounds(
-  grid: CanvasRectangle[][],
-  index: number,
-  cells: number,
-): number {
-  const columns = grid.map((row) => row[0])
-  if (columns.length <= index) {
-    return 0
-  }
-
-  const currentColumn = columns[index]
-
-  if (cells <= 1) {
-    return currentColumn.height
-  }
-
-  function getPadding() {
-    if (grid.length <= 1) {
-      return 0
-    }
-    return grid[1][0].y - (grid[0][0].y + grid[0][0].height)
-  }
-  const padding = getPadding()
-
-  return columns.slice(index + 1, index + cells).reduce((acc, curr) => {
-    return acc + curr.height + padding
-  }, currentColumn.height)
-}
-
 export const GridHelperControls = () => {
   const helperControlsStore = React.useContext(HelperControlsStateContext)
   return (
@@ -2801,50 +2166,3 @@ export const GridHelperControls = () => {
   )
 }
 GridHelperControls.displayName = 'GridHelperControls'
-
-const GridCellOffsetLine = React.memo(
-  (props: { top: number; left: number; size: number; orientation: 'vertical' | 'horizontal' }) => {
-    const colorTheme = useColorTheme()
-
-    return (
-      <div
-        style={{
-          position: 'absolute',
-          top: props.top,
-          left: props.left,
-          width: props.orientation === 'horizontal' ? props.size : 1,
-          height: props.orientation === 'vertical' ? props.size : 1,
-          borderLeft:
-            props.orientation === 'vertical' ? `1px dashed ${colorTheme.primary.value}` : undefined,
-          borderTop:
-            props.orientation === 'horizontal'
-              ? `1px dashed ${colorTheme.primary.value}`
-              : undefined,
-          pointerEvents: 'none',
-        }}
-      />
-    )
-  },
-)
-GridCellOffsetLine.displayName = 'GridCellOffsetLine'
-
-const GridCellOutline = React.memo(
-  (props: { top: number; left: number; width: number; height: number }) => {
-    const colorTheme = useColorTheme()
-
-    return (
-      <div
-        style={{
-          position: 'absolute',
-          top: props.top,
-          left: props.left,
-          width: props.width,
-          height: props.height,
-          border: `1px dashed ${colorTheme.primary.value}`,
-          pointerEvents: 'none',
-        }}
-      />
-    )
-  },
-)
-GridCellOutline.displayName = 'GridCellOutline'
